% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resta.R
\name{resta}
\alias{resta}
\title{Title of the function}
\usage{
resta(a, b)
}
\arguments{
\item{x}{Description of parameter x}

\item{y}{Description of parameter y}
}
\value{
subtraction of x minus y
}
\description{
Description of what the function does.
}
\examples{
resta(1, 2)
}
