% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{suma}
\alias{suma}
\title{Title of the function}
\usage{
suma(a, b)
}
\arguments{
\item{x}{Description of parameter x}

\item{y}{Description of parameter y}
}
\value{
What the function returns
}
\description{
Description of what the function does.
}
\examples{
my_function(1, 2)
}
